requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.8
django-crispy-forms==1.13.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.7.0
python-dotenv==0.19.0
pytz==2021.3
rsa==4.7.2
sqlparse==0.4.2
Telethon==1.23.0
toml==0.10.2
